+++ Service.Manafacturing.Model/ProductionOrder/ProductionOrderDetailModel.cs
@@ -11,8 +11,8 @@ namespace VErp.Services.Manafacturing.Model.ProductionOrder
     public class ProductionOrderDetailOutputModel : ProductionOrderExtraInfo
     {
         public long ProductionOrderId { get; set; }
-        public int? Quantity { get; set; }
-        public int? ReserveQuantity { get; set; }
+        public decimal? Quantity { get; set; }
+        public decimal? ReserveQuantity { get; set; }
         public string Note { get; set; }
         public int Status { get; set; }
     }
@@ -22,8 +22,8 @@ namespace VErp.Services.Manafacturing.Model.ProductionOrder
         public long? ProductionOrderDetailId { get; set; }
         public long ProductionOrderId { get; set; }
         public int? ProductId { get; set; }
-        public int? Quantity { get; set; }
-        public int? ReserveQuantity { get; set; }
+        public decimal? Quantity { get; set; }
+        public decimal? ReserveQuantity { get; set; }
         public string Note { get; set; }
         public long? OrderDetailId { get; set; }
     }
@@ -34,7 +34,7 @@ namespace VErp.Services.Manafacturing.Model.ProductionOrder
         public long? ProductionOrderDetailId { get; set; }
         public long OrderDetailId { get; set; }
         public decimal? OrderQuantity { get; set; }
-        public int? OrderedQuantity { get; set; }
+        public decimal? OrderedQuantity { get; set; }
         public string PartnerId { get; set; }
         public string ProductTitle { get; set; }
         public string OrderCode { get; set; }
